% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaPhiDS.R
\name{alphaPhiDS}
\alias{alphaPhiDS}
\title{Computes the parameters alpha and phi}
\usage{
alphaPhiDS(data, formula, family, clusterID, corstr, startBetas)
}
\arguments{
\item{data}{the input dataframe which contains the variable specified in the 
formula.}

\item{formula}{a regression formula.}

\item{family}{an object of class Family.}

\item{clusterID}{the name of the column that holds the cluster IDs.}

\item{corstr}{the correlation structure.}

\item{startBetas}{a character, the starting values concatenated by comma
because it is not possible to use 'c()' in aggregate functions.}
}
\value{
a list
}
\description{
This function is called by the client function 'ds.gee'
to calculate the parameters alpha and phi.
}
\details{
the parameters are calculated according to the correlation structure.
}
\author{
Gaye, A.; Jones EM.
}
