% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCharacterDS.o.R
\name{asCharacterDS.o}
\alias{asCharacterDS.o}
\title{Coerces an R object into class character}
\usage{
asCharacterDS.o(x.name)
}
\arguments{
\item{x.name}{the name of the input object to be coerced to class
integer. Must be specified in inverted commas. But this argument is
usually specified directly by <x.name> argument of the clientside function
{ds.asCharacter.o}}
}
\value{
the object specified by the <newobj> argument (or its default name
<x.name>.char) which is written to the serverside. For further
details see help on the clientside function {ds.asCharacter.o}
}
\description{
this function is based on the native R function {as.character}
}
\details{
See help for function {as.character} in native R
}
\author{
Amadou Gaye, Paul Burton for DataSHIELD Development Team
}
