% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDataFrameDS.R
\name{asDataFrameDS}
\alias{asDataFrameDS}
\title{asDataFrameDS a serverside assign function called by ds.asDataFrame}
\usage{
asDataFrameDS(x.name)
}
\arguments{
\item{x.name}{the name of the input object to be coerced to class
data.frame. Must be specified in inverted commas. But this argument is
usually specified directly by <x.name> argument of the clientside function
\code{ds.asDataFrame}}
}
\value{
the object specified by the <newobj> argument (or its default name
"asdataframe.newobj") which is written to the serverside. For further
details see help on the clientside function \code{ds.asDataMatrix}
}
\description{
Coerces an R object into a matrix maintaining original
class for all columns in data.frames.
}
\details{
This assign function is based on the native R function \code{data.frame}
}
\author{
Tim Cadman
}
