% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDataMatrixDS.R
\name{asDataMatrixDS}
\alias{asDataMatrixDS}
\title{asDataMatrixDS a serverside assign function called by ds.asDataMatrix}
\usage{
asDataMatrixDS(x.name)
}
\arguments{
\item{x.name}{the name of the input object to be coerced to class
data.matrix. Must be specified in inverted commas. But this argument is
usually specified directly by <x.name> argument of the clientside function
\code{ds.asDataMatrix}}
}
\value{
the object specified by the <newobj> argument (or its default name
"asdatamatrix.newobj") which is written to the serverside. For further
details see help on the clientside function \code{ds.asDataMatrix}
}
\description{
Coerces an R object into a matrix maintaining original
class for all columns in data.frames.
}
\details{
This assign function is based on the native R function \code{data.matrix}
If applied to a data.frame, the native R function \code{as.matrix}
converts all columns into character class. In contrast, if applied to
a data.frame the native R function \code{data.matrix} converts
the data.frame to a matrix but maintains all data columns in their
original class
}
\author{
Paul Burton for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
