% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asListDS.R
\name{asListDS}
\alias{asListDS}
\title{asListDS a serverside assign function called by ds.asList}
\usage{
asListDS(x.name)
}
\arguments{
\item{x.name}{the name of the input object to be coerced to class
data.matrix. Must be specified in inverted commas. But this argument is
usually specified directly by <x.name> argument of the clientside function
\code{ds.asList}}
}
\value{
the input object coerced to a list, which \code{ds.asList} writes to the
serverside. The way that \code{as.list} coerces objects to list depends on the class
of the object
}
\description{
Coerces an R object into a list
}
\details{
This assign function is based on the native R function \code{as.list}
and so additional information can be found in the help for \code{as.list}
}
\author{
Amadou Gaye, Paul Burton for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
