% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMatrixDS.o.R
\name{asMatrixDS.o}
\alias{asMatrixDS.o}
\title{Coerces an R object into a matrix}
\usage{
asMatrixDS.o(x.name)
}
\arguments{
\item{x.name}{the name of the input object to be coerced to class
matrix. Must be specified in inverted commas. But this argument is
usually specified directly by <x.name> argument of the clientside function
{ds.asMatrix.o}}
}
\value{
the object specified by the <newobj> argument (or its default name
<x.name>.mat) which is written to the serverside. For further
details see help on the clientside function {ds.asMatrix.o}
}
\description{
this function is based on the native R function {as.matrix}
}
\details{
See help for function {as.matrix} in native R
}
\author{
Amadou Gaye, Paul Burton for DataSHIELD Development Team
}
