% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucDS.R
\name{aucDS}
\alias{aucDS}
\title{aucDS an aggregate function called by ds.auc}
\usage{
aucDS(pred = pred, y = y)
}
\arguments{
\item{pred}{the name of the vector of the predicted values}

\item{y}{the name of the outcome variable. Note that this variable should include 
the complete cases that are used in the regression model.}
}
\value{
returns the AUC and its standard error
}
\description{
This function calculates the C-statistic or AUC for logistic 
regression models.
}
\details{
The AUC determines the discriminative ability of a model.
}
\author{
Demetris Avraam for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
