% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlotGG_data_TreatmentDS.R
\name{boxPlotGG_data_TreatmentDS}
\alias{boxPlotGG_data_TreatmentDS}
\title{Arrange data frame to pass it to the boxplot function}
\usage{
boxPlotGG_data_TreatmentDS(table.name, variables, group = NULL, group2 = NULL)
}
\arguments{
\item{table.name}{\code{character} Name of a server-side data frame that holds the information to be plotted later}

\item{variables}{\code{character vector} Name of the column(s) of the data frame to include on the boxplot}

\item{group}{\code{character} (default \code{NULL}) Name of the first grouping variable.}

\item{group2}{\code{character} (default \code{NULL}) Name of the second grouping variable.}
}
\value{
\code{data frame} with the following structure: \cr

 Column 'x': Names on the X axis of the boxplot, aka variables to plot \cr
 Column 'value': Values for that variable (raw data of columns rbinded) \cr
 Column 'group': (Optional) Values of the grouping variable \cr
 Column 'group2': (Optional) Values of the second grouping variable \cr
}
\description{
Arrange data frame to pass it to the boxplot function
}
\author{
Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
