\name{cbind}
\alias{cbind}
\title{Take a sequence of vector, matrix or data frames arguments and combine by columns or rows, respectively.}
\usage{
  cbind(...)
}
\arguments{
  \item{...}{vectors or matrices.}
}
\description{
  Take a sequence of vector, matrix or data frames
  arguments and combine by columns or rows, respectively.
}

