% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDS.R
\name{cDS}
\alias{cDS}
\title{Concatenates objects into a vector or list}
\usage{
cDS(x.names)
}
\arguments{
\item{x.names}{a character vector of object names to concatenate.}
}
\value{
a vector or list
}
\description{
This function is similar to the R base function 'c'.
}
\details{
Unlike the R base function 'c' on vector or list of certain 
length are allowed as output
}
\author{
Gaye, A.

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
