% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPermissivePrivacyControlLevel.R
\name{checkPermissivePrivacyControlLevel}
\alias{checkPermissivePrivacyControlLevel}
\title{checkPermissivePrivacyControlLevel}
\usage{
checkPermissivePrivacyControlLevel(privacyControlLevels)
}
\arguments{
\item{privacyControlLevels}{is a vector of strings which contains the privacy control level names which are permitted by the calling method.}
}
\value{
No return value, called for side effects
}
\description{
This server-side function check that the server is running in "permissive" privacy control level.
}
\details{
Tests whether the R option "datashield.privacyControlLevel" is set to "permissive", if it isn't
will cause a call to stop() with the message "BLOCKED: The server is running in 'non-permissive' mode which 
has caused this method to be blocked".
}
\author{
Wheater, Dr SM., DataSHIELD Development Team.
}
