% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDS.R
\name{classDS}
\alias{classDS}
\title{Returns the class of an object}
\usage{
classDS(x)
}
\arguments{
\item{x}{a string character, the name of an object}
}
\value{
the class of the input object
}
\description{
This function is similar to R function \code{class}.
}
\details{
The function returns the class of an object
}
\author{
Stuart Wheater, for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
