% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnamesDS.R
\name{colnamesDS}
\alias{colnamesDS}
\title{Returns the column names of a data frame or matrix}
\usage{
colnamesDS(x)
}
\arguments{
\item{x}{a string character, the name of a dataframe or matrix}
}
\value{
the column names of the input object
}
\description{
This function is similar to R function \code{colnames}.
}
\details{
The function returns the column names of the input dataframe or matrix
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
