% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnClassesDS.R
\name{columnClassesDS}
\alias{columnClassesDS}
\title{Get classes of data frame columns}
\usage{
columnClassesDS(x)
}
\arguments{
\item{x}{\code {character} Name of a data.frame on the server}
}
\value{
\code{named character vector} with the class of each column, the name 
corresponds to the column name
}
\description{
Get classes of data frame columns
}
