\name{complete.cases}
\alias{complete.cases}
\title{Return a logical vector indicating which cases are complete, i.e., have no missing values.}
\usage{
  complete.cases(a, b = NULL)
}
\arguments{
  \item{...}{a sequence of vectors, matrices and data
  frames.}
}
\description{
  Return a logical vector indicating which cases are
  complete, i.e., have no missing values.
}

