% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeDS.R
\name{completeDS}
\alias{completeDS}
\title{DataSHIELD server-side wrapper function for mice::complete}
\usage{
completeDS(mids_obj = NULL, action = 1, include = FALSE)
}
\arguments{
\item{mids_obj}{a character string specifying the name of the mids object containing the 
multiply imputed datasets.}

\item{action}{a numeric value or character string specifying which imputation to extract. 
The default is 1, which returns the first completed dataset. Other possibilities are 
"long", "broad", "repeated". See mice::complete documentation for details.}

\item{include}{a logical value indicating whether the original data should be included. 
Only relevant when action is "long" or "repeated". Default is FALSE.}
}
\value{
a data.frame containing the completed data as specified by the action parameter.
}
\description{
This function is a wrapper for the complete function from the mice R package.
It extracts the completed data from a mids object containing multiply imputed datasets.
}
\details{
For additional details see the help header of complete function in native R mice package.
}
\author{
DataSHIELD Development Team
}
