\name{contour.plot.1}
\alias{contour.plot.1}
\title{Computes a grid-density object required to generate a contour plot.}
\usage{
  contour.plot.1(xvect, yvect)
}
\arguments{
  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\value{
  a heatmap plot.
}
\description{
  Computes a grid-density object required to generate a
  contour plot.
}
\author{
  Burton, P.
}

