% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.zphSLMADS.R
\name{cox.zphSLMADS}
\alias{cox.zphSLMADS}
\title{Tests the proportional hazards assumption of a Cox proportional
 hazards model that has been fit and saved serverside.}
\usage{
cox.zphSLMADS(fit = NULL, dataName = NULL)
}
\arguments{
\item{fit}{character string specifying name of fit Cox proportional 
hazards model saved in the server-side.}

\item{dataName}{character string of name of data frame}
}
\value{
diagnostics for the Cox proportional hazards from the server side environment.
}
\description{
Tests the proportional hazards assumption of a 
Cox proportional hazards that has been fit and saved on the
server side environment.
}
\details{
Serverside aggregate function {cox.zphSLMADS} called by clientside function.
{ds.cox.zphSLMA}.
returns diagnostics for the test of proportional hazards assumptions
from a Cox proportional hazards model.
This request is not disclosive as it only returns summary statistics.
For further details see help for {ds.cox.zphSLMA} function.
}
\author{
Soumya Banerjee and Tom Bishop (2020).
}
