% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxPHMDS.R
\name{coxPHMDS}
\alias{coxPHMDS}
\title{coxPHMDS is the serviceside function for Cox Proposional Hazard Model}
\usage{
coxPHMDS(
  data.name,
  event.variable,
  term.variables,
  strata.variables,
  start.variable = NULL,
  stop.variable = NULL,
  period.variable = NULL
)
}
\arguments{
\item{data.name}{is a character string contained the name of the data variable}

\item{event.variable}{is a character string contains the name of the event field in the data}

\item{term.variables}{is a vector of character string contains the names of the term fields in the data}

\item{strata.variables}{is a vector of character string contains the names of the terms fields in the data, which are strata}

\item{start.variable}{is a character string contains the name of the start field in the data}

\item{stop.variable}{is a character string contains the name of the stop field in the data}

\item{period.variable}{is a character string contains the name of the period field in the data}
}
\value{
the results of the Cox Proposional Hazard Model
}
\description{
creates Cox Proposional Hazard Model
}
\details{
this serverside function returns the
}
\author{
Stuart Wheater for DataSHIELD Development Team, Jun 2020
}
