% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameFillDS.R
\name{dataFrameFillDS}
\alias{dataFrameFillDS}
\title{dataFrameFillDS}
\usage{
dataFrameFillDS(
  df.name,
  allNames.transmit,
  class.vect.transmit,
  levels.vec.transmit
)
}
\arguments{
\item{df.name}{a character string representing the name of the input data frame that will be
filled with extra columns with missing values if a number of variables is missing from it
compared to the data frames of the other studies used in the analysis.}

\item{allNames.transmit}{unique names of all the variables that are included in the input 
data frames from all the used datasources.}

\item{class.vect.transmit}{the classes of all the variables that are included in the vector 
\code{allNames.transmit}.}

\item{levels.vec.transmit}{the levels of all factor variables.
The classes supported are 'numeric', 'integer', 'character', 'factor' and 'logical'.}
}
\value{
Nothing is returned to the client. The generated object is written to the serverside.
}
\description{
An assign function called by the clientside ds.dataFrameFill function.
}
\details{
This function checks if each study has all the variables compared to the other studies
in the analysis. If a study does not have some of the variables, the function generates those
variables as vectors of missing values and combines them as columns to the input data frame. 
Then, the "complete" in terms of the columns dataframe is saved in each server with a name
specified by the argument \code{newobj} on the clientside.
}
\author{
Demetris Avraam for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
