% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deltaDS.R
\name{deltaDS}
\alias{deltaDS}
\title{Calculates the difference between rows for a target column in grouped data}
\usage{
deltaDS(input, id_col, target_col)
}
\arguments{
\item{input}{a character, the name of the data frame to process}

\item{id_col}{a character, the name of the field in the data frame that groups
the data}

\item{target_col}{a character, the name of the field in the data frame}
}
\value{
a new vector containing the deltas
}
\description{
This function calculates the difference between rows for a target
column in grouped data. For non-grouped data it would be possible to use rowShift
}
\details{
This function is used when the analyst needs to calculate the difference
between rows for a column. The data should be grouped by an identifying column:
 for non-grouped data it is possible to use the rowShift function and a subtraction
 to perform the same operation
}
\author{
Bishop, T.
}

