% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimDS.o.R
\name{dimDS.o}
\alias{dimDS.o}
\title{Returns the dimension of a data frame or matrix}
\usage{
dimDS.o(x)
}
\arguments{
\item{x}{a string character, the name of a dataframe or matrix}
}
\value{
the dimension of the input object
}
\description{
This function is similar to R function \code{dim}.
}
\details{
The function returns the dimension of the input dataframe or matrix
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
