% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsDS.R
\name{existsDS}
\alias{existsDS}
\title{Is an object exists defined with an expected internal type?}
\usage{
existsDS(
  variable.name = NULL,
  environment.name = ".GlobalEnv",
  class.type = NULL
)
}
\arguments{
\item{variable.name}{A character value that state the variable (R object) to check}

\item{environment.name}{A character value stating the name of an environment}

\item{class.type}{A character value stating the R internal type. Correct values:
\itemize{
\item "\code{\link{character}}"
\item "\code{\link{complex}}"
\item "\code{\link{double}}"
\item "\code{\link{expression}}"
\item "\code{\link{integer}}"
\item "\code{\link{list}}"
\item "\code{\link{logical}}"
\item "\code{\link{numeric}}"
\item "\code{\link{single}}"
\item "\code{\link{raw}}"
\item "\code{\link{vector}}"
\item "\code{\link{S4}}"
\item "\code{\link{NULL}}"
\item "\code{\link{function}}"
\item "\code{externalptr}"
\item "\code{\link{environment}}"
}}
}
\value{
\itemize{
\item \code{TRUE} if an object is found in a specific environment and it is of a specific type
\item \code{FALSE}, if the object does not exists in an environment or it is not made an expected class
}
}
\description{
This server function indicates whether a variable or R object exists in a specific environment. It also verifies
the class of the object correspond to an expected class.
}
\details{
This function uses the \code{\link{new.env}} R function to create an  environment. The default environment should is set to ".GlobalEnv".
However, any environments available on the search path can be specified; use \code{\link{search}} function to identify them.

The function \code{\link{class}} verifies the R object has an expected class.

Type: aggregate server function
}
\seealso{
\code{\link{exists}}, \code{\link{typeof}}, \code{\link{class}}, \url{https://stat.ethz.ch/R-manual/R-devel/library/methods/html/BasicClasses.html}
}
\author{
P.Ryser-Welch, Newcastle University, DataSHIELD team
}
