% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDS.R
\name{expDS}
\alias{expDS}
\title{Computes the exponential values of the input variable}
\usage{
expDS(x)
}
\arguments{
\item{x}{a string character, the name of a numeric or integer vector}
}
\value{
the object specified by the \code{newobj} argument
of \code{ds.exp} (or default name \code{exp.newobj})
which is written to the serverside. The output object is of class numeric.
}
\description{
This function is similar to R function \code{exp}.
}
\details{
The function computes the exponential values of an input numeric
or integer vector.
}
\author{
DataSHIELD Development Team
}
