% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseDfDS.R
\name{fixClassDS}
\alias{fixClassDS}
\title{Change Class of Target Variables in a Data Frame}
\usage{
fixClassDS(df.name, target_vars, target_class)
}
\arguments{
\item{df.name}{A string representing the name of the data frame.}

\item{target_vars}{A character vector specifying the columns to be modified.}

\item{target_class}{A character vector specifying the new classes for each column (1 = factor,
2 = integer, 3 = numeric, 4 = character, 5 = logical).}
}
\value{
A modified data frame with the specified columns converted to the target classes.
}
\description{
Change Class of Target Variables in a Data Frame
}
