% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseDfDS.R
\name{fixColsDS}
\alias{fixColsDS}
\title{Add Missing Columns with NA Values}
\usage{
fixColsDS(.data, cols)
}
\arguments{
\item{.data}{A string representing the name of the data frame.}

\item{cols}{A character vector specifying the columns to be added if missing.}
}
\value{
A modified data frame with missing columns added and filled with NA.
}
\description{
Add Missing Columns with NA Values
}
