% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseDfDS.R
\name{fixLevelsDS}
\alias{fixLevelsDS}
\title{Set Factor Levels for Specific Columns in a Data Frame}
\usage{
fixLevelsDS(df.name, vars, levels)
}
\arguments{
\item{df.name}{A string representing the name of the data frame to modify.}

\item{vars}{A character vector specifying the columns to be modified.}

\item{levels}{A named list where each element contains the levels for the corresponding factor variable.}
}
\value{
A modified data frame with the specified columns converted to factors with the provided levels.
}
\description{
Set Factor Levels for Specific Columns in a Data Frame
}
