% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseDfDS.R
\name{getAllLevelsDS}
\alias{getAllLevelsDS}
\title{Retrieve Factor Levels for Specific Columns}
\usage{
getAllLevelsDS(df.name, factor_vars)
}
\arguments{
\item{df.name}{A string representing the name of the data frame.}

\item{factor_vars}{A character vector specifying the factor columns.}
}
\value{
A list of factor levels for the specified columns.
}
\description{
Retrieve Factor Levels for Specific Columns
}
