% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvironmentDS.R
\name{getEnvironmentDS}
\alias{getEnvironmentDS}
\title{retrieves an environment from its name.}
\usage{
getEnvironmentDS(environment.name = NULL, package.name = ".GlobalEnv")
}
\arguments{
\item{environment.name}{the name of an environment available on a DataSHIELD server.}

\item{package.name}{the name of the package as appearing in \code{\link{search}}}
}
\value{
\itemize{
\item  an empty enviroment if no environment is found.
\item  the reference to the environment itself.
}
}
\description{
retrieves an environment from its name. It verifies the environment exists within the specified parent environment, before returning the environment itself.
}
\details{
The helper function verifies a environment exists on the search path.By default, the parent environment used is the ".GlobalEnv".

The helper function verifies a environment exists, within another environment  on a dataSHIELD server. By defautl the parent environment used is the
\code{globalenv()}. For testing and other purposes, other parent environment can be specify
}
\author{
P.Ryser-Welch, Newcastle University, DataSHIELD team
}
