% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSLMADS1.R
\name{glmSLMADS1}
\alias{glmSLMADS1}
\title{Fit a Generalized Linear Model (GLM) with pooling via Study Level Meta-Analysis (SLMA)}
\usage{
glmSLMADS1(formula, family, weights, offset, data)
}
\arguments{
\item{formula}{a glm formula, specified in call to ds.glmSLMA}

\item{family}{a glm family, specified in call to ds.glmSLMA}

\item{weights}{a character string specifying a variable to be used as regression weights.
Specified in call to ds.glmSLMA. Specified in call to ds.glmSLMA.}

\item{offset}{a character string specifying a variable to be used as an offset.
Specified in call to ds.glmSLMA.}

\item{data}{a character string specifying the name of a data.frame
holding the data for the model. Specified as dataName in call to ds.glmSLMA.}
}
\value{
assesses and returns information about failure to pass disclosure traps
such as test of model complexity (saturation).
For more detailed information see help for ds.glmSLMA.
}
\description{
Fits a generalized linear model (GLM) on data from single or multiple sources
with pooled co-analysis across studies being based on SLMA (Study Level Meta Analysis).
}
\details{
glmSLMADS.assign is an aggregate function called by clientside function ds.glmSLMA.
ds.glmSLMA also calls another aggregate function glmSLMADS2
and an assign function glmSLMADS.assign
For more detailed information see help for ds.glmSLMA.
}
\author{
Paul Burton for DataSHIELD Development Team (14/7/20)

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
