% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSLMADS2.o.R
\name{glmSLMADS2.o}
\alias{glmSLMADS2.o}
\title{glmSLMADS2.o}
\usage{
glmSLMADS2.o(formula, family, offset, weights, dataName)
}
\arguments{
\item{formula}{a glm() formula consistent with R syntax eg U~x+y+Z to regress
variables U on x,y and Z}

\item{family}{a glm() family consistent with R syntax eg "gaussian", "poisson",
"binomial"}

\item{offset}{an optional variable providing a regression offset}

\item{weights}{an optional variable providing regression weights}

\item{dataName}{an optional character string specifying a data.frame object holding
the data to be analysed under the specified model.}
}
\value{
model components:- glmDSDLMA2.o returns key components of model fit
from each study including parameter estimates and standard errors which
are then processed and reported by ds.glmSLMA.o potentially including
random effects meta-analysis using the metafor package if requested
in the call to ds.glmSLMA.o
}
\description{
This is the second serverside function called by ds.glmSLMA.o.
}
\details{
It is an
aggregation function that fits the generalized linear model (glm) specified
in the call to ds.glmSLMA.o. The model is first constructed
and disclosure checked by glmDS1.o (the same serverside function called first
by ds.glm.o as well as by ds.glmSLMA.o). It is then fitted to convergence
on each study separately using glmSLMADS2.o to return parameter estimates
and standard errors to the client. These can then be pooled using random
effects meta-analysis (eg under metafor). This mode of model fitting may
reasonably be called study level meta-analysis (SLMA) although the analysis
is based on estimates and standard errors derived from direct analysis of
the individual level data in each study rather than from published study
summaries. It may also be contrasted to the approach adopted by ds.glm.o
via glmDS2.o which fits the glm iteratively on all studies simultaneously
updating the parameter estimates across all studies simultaneously as the
model converges. This latter approach may reasonably be called IPD
(individual patient data meta-analysis) and it is mathematically equivalent
to placing all data from all studies in one data warehouse and analysing
them using the standard glm() function in native R (with appropriate
study-specific centre effects). Both of these approaches have strengths
and weaknesses and this is why DataSHIELD supports both methods.  
For more details please see the extensive headers for ds.glm.o and
ds.glmSLMA.o.
}
\author{
Burton PR
}
