\name{histogram.1}
\alias{histogram.1}
\title{Computes a histogram of the given data values without plotting.}
\usage{
  histogram.1(a)
}
\arguments{
  \item{a}{a vector of values for which the histogram is
  desired.}
}
\description{
  Computes a histogram of the given data values without
  plotting.
}

