% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNumVariableValidDS.R
\name{isNumVariableValidDS}
\alias{isNumVariableValidDS}
\title{checks if a variable or a column of data.frame exists on a DataSHIELD server of type numeric}
\usage{
isNumVariableValidDS(x = "")
}
\arguments{
\item{x}{the variable name}
}
\value{
TRUE - if a numerical variable (or column of a dataframe exists). Otherwise, FALSE
}
\description{
checks if a variable or a column of data.frame exists on a DataSHIELD server of type numeric
}
\keyword{server.function}
