% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidDS.R
\name{isValidDS}
\alias{isValidDS}
\title{Checks if an input is valid}
\usage{
isValidDS(x)
}
\arguments{
\item{x}{a character string, the name of a vector (numeric, integer, factor, character,
logical), data.frame or matrix}
}
\value{
a list with \code{valid}, a boolean that is TRUE if the input is valid or FALSE
if not, and \code{class}, the class of the input object for client-side consistency
checking
}
\description{
Tells if an object on the server side is valid.
}
\details{
This function checks if an object is valid.
}
\author{
Gaye, A.

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
