% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosisDS1.R
\name{kurtosisDS1}
\alias{kurtosisDS1}
\title{Calculates the kurtosis of a numeric variable}
\usage{
kurtosisDS1(x, method)
}
\arguments{
\item{x}{a string character, the name of a numeric variable.}

\item{method}{an integer between 1 and 3 selecting one of the algorithms for computing kurtosis
detailed in the headers of the client-side \code{ds.kurtosis} function.}
}
\value{
a list including the kurtosis of the input numeric variable, the number of valid observations and
the study-side validity message.
}
\description{
This function calculates the kurtosis of a numeric variable for each study separately.
}
\details{
The function calculates the kurtosis of an input variable x with three different methods. 
The method is specified by the argument \code{method} in the client-side \code{ds.kurtosis} function.
}
\author{
Demetris Avraam, for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
