% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthDS.o.R
\name{lengthDS.o}
\alias{lengthDS.o}
\title{Returns the length of a vector or list}
\usage{
lengthDS.o(x)
}
\arguments{
\item{x}{a string character, the name of a vector or list}
}
\value{
a numeric, the number of elements of the input vector or list.
}
\description{
This function is similar to R function \code{length}.
}
\details{
The function returns the length of the input vector or list.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
