% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisDS3.R
\name{lexisDS3}
\alias{lexisDS3}
\title{@title
lexisDS3}
\usage{
lexisDS3()
}
\value{
Data frame with `messageobj` object
}
\description{
The third serverside function called by ds.lexis.
}
\details{
This is an assign function that simplifies the
returned output from ds.lexis. Specifically, without lexisDS3 the output consists of a table within
a list, but lexisDS3 converts this directly into a dataframe.
For more details see the extensive header for ds.lexis.
}
