% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexishelper1.R
\name{lexishelper1}
\alias{lexishelper1}
\title{Computes time breaks given the exit time}
\usage{
lexishelper1(stopTime, interval)
}
\arguments{
\item{stopTime}{a numeric, the exit time time.}

\item{interval}{a numeric vector of one or more elements which give the sought
interval width.}
}
\value{
a list which contains the break points for each subject.
}
\description{
This is an internal function required by the client.
function \code{ds.lexis} to get the interval break points for each subject.
}
\details{
The strategy used in this function consists of walking through the 
specified interval width and creating interval. Each time an interval is 
created the exit time is reduced by the same magnitude of time. When the remaining
exit time is smaller than the next interval to create to process stops and 
that last interval is then the interval in which the subject fails. The remaining 
exit time is in fact the survival time in the last interval where the subject failed.
}
\author{
Gaye, A.
}
\keyword{internal}
