% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexishelper2.R
\name{lexishelper2}
\alias{lexishelper2}
\title{Computes time periods of each subject}
\usage{
lexishelper2(dt, ids, exits, interval)
}
\arguments{
\item{dt}{the dataset that contains the original data.}

\item{ids}{a string character, the name of column that holds the individuals IDs.}

\item{interval}{a numeric vector, the sought interval widths.}

\item{exists, }{a string character, the name of column that holds the exit times.}
}
\value{
a list which contains the individuals time periods.
}
\description{
This is an internal function required by the client.
function \code{ds.lexis} to get the time periods for each subject.
}
\details{
If the exit time information is missing one single break
is returned with the a missing value.
}
\author{
Gaye, A.
}
\keyword{internal}
