% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDS.R
\name{listDS}
\alias{listDS}
\title{Coerce objects into a list}
\usage{
listDS(x.names = NULL, eltnames = NULL)
}
\arguments{
\item{x.names}{a character vector of object names to coerce into a list.}

\item{eltnames}{a character vector, the names of the elements in the list.}
}
\value{
a list
}
\description{
this function is similar to R function 'list'
}
\details{
Unlike the R function 'list' it takes also a vector of characters,
the names of the elements in the output list.
}
\author{
Gaye, A.

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
