% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listNaDS.R
\name{listNaDS}
\alias{listNaDS}
\title{Indicates the position of missing values in a vector}
\usage{
listNaDS(xvect)
}
\arguments{
\item{xvect}{a character, the name of the vector to process.}
}
\value{
a new vector indicating where missing values are found
}
\description{
This function identifies the position of missing values
}
\details{


This function is used when the analyst needs to know where missing values
are located. This is useful when using methods such as Missing Indicator.
}
\author{
Bishop, T.
}

