% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSLMADS.assign.R
\name{lmerSLMADS.assign}
\alias{lmerSLMADS.assign}
\title{Fitting linear mixed effect models - serverside function}
\usage{
lmerSLMADS.assign(
  formula,
  offset,
  weights,
  dataName,
  REML = TRUE,
  control_type,
  control_value.transmit,
  optimizer,
  verbose = 0
)
}
\arguments{
\item{formula}{see help for ds.lmerSLMA}

\item{offset}{see help for ds.lmerSLMA}

\item{weights}{see help for ds.lmerSLMA}

\item{dataName}{see help for ds.lmerSLMA}

\item{REML}{see help for ds.lmerSLMA}

\item{control_type}{see help for ds.lmerSLMA}

\item{control_value.transmit}{see help for argument <control_value> for
function ds.lmerSLMA}

\item{optimizer}{see help for ds.lmerSLMA}

\item{verbose}{see help for ds.lmerSLMA}
}
\value{
writes lmerMod object summarising the fitted model to the serverside.
For more detailed information see help for ds.lmerSLMA.
}
\description{
lmerSLMADS.assign is the same as lmerSLMADS2 which fits a linear
mixed effects model (lme) per study and saves the outcomes in each study
}
\details{
lmerSLMADS.assign is a serverside function called by ds.lmerSLMA on the clientside.
The analytic work engine is the lmer function in R which sits in the lme4 package.
lmerSLMADS.assign fits a linear mixed effects model (lme) including both fixed and random
effects - on data from each single data source and saves the regression outcomes on the serverside.
}
\author{
TDemetris Avraam for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
