% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDS.R
\name{logDS}
\alias{logDS}
\title{Computes the logarithm values of the input variable}
\usage{
logDS(x, base = exp(1))
}
\arguments{
\item{x}{a string character, the name of a numeric or integer vector}

\item{base}{a positive number, the base for which logarithms are computed.
Default \code{exp(1)}.}
}
\value{
the object specified by the \code{newobj} argument
of \code{ds.log} (or default name \code{log.newobj})
which is written to the serverside. The output object is of class numeric.
}
\description{
This function is similar to R function \code{log}.
}
\details{
The function computes the logarithm values of an input numeric
or integer vector. By default natural logarithms are computed.
}
\author{
DataSHIELD Development Team
}
