% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDetDS1.R
\name{matrixDetDS1}
\alias{matrixDetDS1}
\title{matrixDetDS aggregate function called by ds.matrixDet.report}
\usage{
matrixDetDS1(M1.name = NULL, logarithm)
}
\arguments{
\item{M1.name}{A character string specifying the name of the matrix for which
determinant to be calculated}

\item{logarithm}{logical. Default is FALSE, which returns the
determinant itself, TRUE returns the logarithm of the modulus of the determinant.}
}
\value{
Output is the determinant of the matrix identified by argument <M1>
which is returned to the clientside. For more details see help for ds.matrixDet
}
\description{
Calculates the determinant of a square matrix A and returns
the output to the clientside
}
\details{
Calculates the determinant of a square matrix (for additional
information see help for \code{det} function in native R). This operation is only
possible if the number of columns and rows of A are the same.
}
\author{
Paul Burton for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
