% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDiagDS.R
\name{matrixDiagDS}
\alias{matrixDiagDS}
\title{matrixDiagDS assign function called by ds.matrixDiag}
\usage{
matrixDiagDS(x1.transmit, aim, nrows.transmit)
}
\arguments{
\item{x1.transmit}{identifies the input matrix or vector. Fully
specified by <x1> argument of \code{ds.matrixDiag}. For more details
see help for \code{ds.matrixDiag}.}

\item{aim}{a character string specifying what behaviour is required
of the function. Fully specified by <aim> argument of \code{ds.matrixDiag}.
For more details see help for \code{ds.matrixDiag}.}

\item{nrows.transmit}{a scalar value forcing the number of rows and
columns in an output matrix.Fully specified by <nrows.scalar>
argument of \code{ds.matrixDiag}.
For more details see help for \code{ds.matrixDiag}.}
}
\value{
Output is the matrix or vector specified by the <newobj> argument
(or default name diag_<x1>) which is written to the serverside.
For more details see help for \code{ds.matrixDiag}.
}
\description{
Extracts the diagonal vector from a square matrix A or
creates a diagonal matrix A based on a vector or a scalar value and
writes the output to the serverside
}
\details{
For details see help for function \code{ds.matrixDiag}.
}
\author{
Paul Burton for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
