% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDimnamesDS.R
\name{matrixDimnamesDS}
\alias{matrixDimnamesDS}
\title{matrixDimnamesDS assign function called by ds.matrixDimnames}
\usage{
matrixDimnamesDS(M1.name = NULL, dimnames)
}
\arguments{
\item{M1.name}{Specifies the name of the serverside matrix to which
dimnames are to be added. Fully specified by <M1> argument of
function \code{ds.matrixDimnames}. For more details
see help for \code{ds.matrixDimnames}.}

\item{dimnames}{A dimnames attribute for the matrix: NULL or a list of
length 2 giving the row and column names respectively.
Fully specified by <dimnames> argument of
function \code{ds.matrixDimnames}. For more details
see help for \code{ds.matrixDimnames}.}
}
\value{
Output is the serverside matrix specified by the <newobj> argument
(or default name diag_<x1>) with specified dimnames (row and column
names) which is written to the serverside.
}
\description{
Adds dimnames (row names, column names or both) to
a matrix on the serverside.
}
\details{
Adds dimnames (row names, column names or both) to
a matrix on the serverside. Similar to the \code{dimnames} function
in native R. For more details see help for
function \code{ds.matrixDimnames}
}
\author{
Paul Burton for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
