% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSdGpDS.R
\name{meanSdGpDS}
\alias{meanSdGpDS}
\title{MeanSdGpDS}
\usage{
meanSdGpDS(x, index)
}
\arguments{
\item{x}{a client-side supplied character string identifying the variable for which
means/SDs are to be calculated}

\item{index}{a client-side supplied character string identifying the factor across
which means/SDs are to be calculated}
}
\value{
List with results from the group statistics
}
\description{
Server-side function called by ds.meanSdGp
}
\details{
Computes the mean and standard deviation across groups defined by one
factor
}
\author{
Burton PR

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
