% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memUsedDS.R
\name{memUsedDS}
\alias{memUsedDS}
\title{Returns memory used in R session}
\usage{
memUsedDS()
}
\value{
The memory used in the R session
}
\description{
This function is similar to R function \code{pryr::mem_used}.
}
\author{
Tim Cadman
}
