% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numNaDS.R
\name{numNaDS}
\alias{numNaDS}
\title{Counts the number of missing values}
\usage{
numNaDS(x)
}
\arguments{
\item{x}{a character string, the name of a server-side vector}
}
\value{
an integer, the number of missing values
}
\description{
this function just counts the number of missing entries
in a vector.
}
\author{
Gaye, A.
}
