\name{odds.ratio}
\alias{odds.ratio}
\title{Calculates odds ratio and its 95 percent confidence interval}
\usage{
  odds.ratio(xvect, yvect, alpha = 0.05)
}
\arguments{
  \item{xvect}{a factor vector}

  \item{yvect}{a factor vector}

  \item{alpha}{level of statistical significance}
}
\description{
  Calculates odds ratio and its 95 percent confidence
  interval
}

