% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-otel.R
\name{otel_add_attributes}
\alias{otel_add_attributes}
\title{Add list of attributes to \link[otel]{otel_span} object}
\usage{
otel_add_attributes(span, attrs)
}
\arguments{
\item{span}{\link[otel]{otel_span} object.}

\item{attrs}{list with named attributes.}
}
\value{
\code{span} with added attributes, \code{attrs}.
}
\description{
Add list of attributes to \link[otel]{otel_span} object
}
\keyword{internal}
