% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDS.R
\name{predictDS}
\alias{predictDS}
\title{predictDS}
\usage{
predictDS(
  newdataname,
  traindataname,
  type = c("response", "link"),
  na.action = "na.pass"
)
}
\arguments{
\item{newdataname}{A character string specifying the name of the new dataset to be used for predictions.}

\item{traindataname}{A character string specifying the name of the dataset used for model training.}

\item{type}{A character string specifying the type of prediction. Options are \code{"response"} or \code{"link"}.}

\item{na.action}{A character string to specify the action to take if missing values are present. Default is \code{"na.pass"}.}
}
\value{
a numeric vector containing the predicted values
}
\description{
Generates server-side predictions using the client-side output from \code{ds.glm}.
}
\details{
This function uses the components supplied by the client-side function (coefficients, family, formula, 
and any categorical variables) to generate predictions on the server. To use the base R \code{predict()} function, 
a "dummy" glm object is created using the same model formula, family, and link function as the original model. 
The dummy model's coefficients are then replaced with the client-side coefficient estimates.

To avoid mismatches between the factors used in the original glm and those in the dummy glm, the categorical 
variables saved by the client-side function are applied to the newdata.

For intercept-only models, the function simply returns a vector of predicted values equal to the model intercept, with the appropriate length 
based on the row length of \code{newdataname}.
}
\author{
Zulal Bekerecioglu
}
