\name{recodelevelsDS}
\alias{recodelevelsDS}
\title{Recodes the levels of a categorical variables}
\usage{
recodelevelsDS(xvect = NULL, classes = NULL)
}
\arguments{
  \item{xvect}{a factor vector}

  \item{classes}{a character vector the levels of the newt
  factor vector}
}
\value{
a factor vector with the new levels
}
\description{
The functions uses the input factor and generates a new
factor with new levels.
}
\details{

}
\author{
Gaye, A.
}

