% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDS.R
\name{removeDS}
\alias{removeDS}
\title{removeDS  - destroys an environment on the server}
\usage{
removeDS(
  variable.name = NULL,
  environment.name = "",
  class.type = NULL,
  package = sys.frame()
)
}
\arguments{
\item{variable.name}{A character value that state the variable (R object) to check}

\item{environment.name}{A character value stating the name of an environment. (1) It could  be the name of the package as appearing in \code{\link{search}}.
It could also be the name of child environment created in the global enviromment or an package environment.}

\item{class.type}{A character value stating the R internal type. Correct values:
\itemize{
\item "\code{\link{character}}"
\item "\code{\link{complex}}"
\item "\code{\link{double}}"
\item "\code{\link{expression}}"
\item "\code{\link{integer}}"
\item "\code{\link{list}}"
\item "\code{\link{matrix}}"
\item "\code{\link{logical}}"
\item "\code{\link{numeric}}"
\item "\code{\link{single}}"
\item "\code{\link{raw}}"
\item "\code{\link{vector}}"
\item  "\code{\link{S4}}"
\item "\code{\link{NULL}}"
\item "\code{\link{function}}"
\item "\code{\link{externalptr}}"
\item "\code{\link{environment}}"
}
#'@param package.name the name of the package as appearing in \code{\link{search}} (optional)}
}
\value{
\itemize{
\item TRUE  The environment has been destroyed successfully
\item FALSE No environment has been destroyed
}
}
\description{
Remove objects (variables) of a specific data type on a DataSHIEL server
}
\details{
This server

TYPE: AGGREGATE SERVER FUNCTION
}
\author{
P.Ryser-Welch, Newcastle University, DataSHIELD team
}
