% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDS.R
\name{roundDS}
\alias{roundDS}
\title{roundDS}
\usage{
roundDS(x, type, digits, add.column, newobj)
}
\arguments{
\item{x}{Character vector specifying the server-side object(s). For data-frame columns, use the format \code{df$column}.}

\item{type}{Character string specifying the operation: \code{"round"}, \code{"ceiling"}, \code{"floor"},
\code{trunc}, or \code{"signif"}.}

\item{digits}{Number of digits to be used in arguments \code{"round"} and \code{"signif"}.}

\item{add.column}{Logical. If \code{FALSE}, the result is created as a new server-side object; 
if \code{TRUE}, the result is added as a new column in the existing data-frame. Default is \code{FALSE}.}

\item{newobj}{Character string for the name of the object that will be created on the server. Default is \code{"rounding.result"}.}
}
\value{
the created numeric vector or the updated dataframe with the added column
}
\description{
Generates objects from a server-side object, which can be either a vector or 
a data-frame column. Supports five operations:  
1. (\code{round})  
2. (\code{ceiling})  
3. (\code{floor})
4. (\code{trunc})
5. (\code{signif})
where each function in baseR is applied on the server side to the specified object.
}
\details{
Note: \code{add.column = TRUE} is only valid for data-frame inputs.
}
\author{
Zulal Bekerecioglu
}
