\name{rowcolDel.ds}
\alias{rowcolDel.ds}
\title{Deletes rows or columns from a dataframe}
\usage{
  rowcolDel.ds(dataset, dimension, indx)
}
\arguments{
  \item{dataset}{a dataframe}

  \item{dimension}{an integer 1 to delete rows and 2 to
  delete columns}

  \item{indx}{a numeric vector that represents the rows or
  columns to remove}
}
\value{
  a dataframe
}
\description{
  Deletes rows or columns from a dataframe
}
\author{
  Gaye, A.
}

