% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowShiftDS.R
\name{rowShiftDS}
\alias{rowShiftDS}
\title{Generates a new column shifted up or down}
\usage{
rowShiftDS(xvect, shiftLen = 1L)
}
\arguments{
\item{xvect}{a character, the name of the vector to process.}

\item{shiftLen}{a positive or negative integer specifying the number of rows to
shift up or down by.}
}
\value{
a new vector with shifted values
}
\description{
This function shifts a vector up or down by a number of rows specified
}
\details{


This function is used when the analyst needs to generate a new column
by shifting an existing column up or down by a number of rows. This is similar to a
lag function in SAS. A negative integer shifts the column down, a positive number
shifts the column up.
}
\author{
Bishop, T.
}

