% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleDS.R
\name{scaleDS}
\alias{scaleDS}
\title{scaleDS}
\usage{
scaleDS(x = NULL, newobj = NULL, add.column = NULL)
}
\arguments{
\item{x}{Character string specifying the server-side vector For data-frame columns, use the format \code{df$column}.}

\item{newobj}{Character string for the name of the object that will be created on the server. Default is \code{"scaled.data"}.}

\item{add.column}{Logical. If \code{FALSE}, the result is created as a new server-side object; 
if \code{TRUE}, the result is added as a new column in the existing data-frame. Default is \code{FALSE}.}
}
\value{
the created numeric vector or the updated dataframe with the added column
}
\description{
Generates scaled objects using a server-side object, which can be either a vector or 
a data-frame column.
}
\details{
Note: \code{add.column = TRUE} is only valid for data-frame inputs.
}
\author{
Zulal Bekerecioglu
}
