\name{setFilterDS}
\alias{setFilterDS}
\title{Sets the privacy level}
\usage{
setFilterDS(x = 5)
}
\arguments{
  \item{x}{a dummy argument}
}
\value{
an integer between 1 and 5
}
\description{
This is an internal function that set the number of
observation that are allowed in any table or vector
}
\details{


The function is called by server side functios
}
\author{
Gaye, A.
}
\keyword{internal}

