% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setRowNameDS.R
\name{setRowNamesDS}
\alias{setRowNamesDS}
\title{Changes the names of rows of a data frame}
\usage{
setRowNamesDS(x, names)
}
\arguments{
\item{x}{a character corresponding the name of a data frame on the server side}

\item{names}{a vector of characters equal in length to the number of rows in x}
}
\value{
a numeric, the statistical mean
}
\description{
Changes the names of rows of a data frame
}
\details{


Needs sorting out so that there is a function to set the name and a function
to get the name
}
\author{
Bishop, T.
}

