% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByClassHelper1.R
\name{subsetByClassHelper1}
\alias{subsetByClassHelper1}
\title{generates subsets vectors from a factor vector}
\usage{
subsetByClassHelper1(xvect = NULL, xname = NULL, filter = NULL)
}
\arguments{
\item{xvect}{a vector of type factor.}

\item{xname}{the name of the vector.}

\item{nfilter}{the minimum number observation (i.e. rows) that are allowed.}
}
\value{
a list which contains the subsets.
}
\description{
This is an internal function called by the function 'subsetByClassDS'.
}
\details{
The function generates subsets if the input of 'subsetByClassDS' is a factor vector.
}
\author{
Gaye, A.
}
\keyword{internal}
