% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByClassHelper2.R
\name{subsetByClassHelper2}
\alias{subsetByClassHelper2}
\title{generates subset tables from a data frame}
\usage{
subsetByClassHelper2(df = NULL, iter = NULL, filter = NULL)
}
\arguments{
\item{df}{a data frame.}

\item{iter}{the indices of columns to loop trough.}

\item{filter}{the minimum number of observations (i.e. rows) that are allowed.}
}
\value{
a list which contains the subsets, their names and an integer that indicates how many columns were
not factors.
}
\description{
This is an internal function called by the function 'subsetByClassDS'
}
\details{
The function generates subsets if the input of 'subsetByClassDS' is a data frame 
and if the number variables(columns) to subset by are greater than 1; i.e. this 
function is called if the user specified more than one variable or no variable to subset by
(if no variables are specified the function 'subsetByClassDS' produces a subset for each category
in each variable).
}
\author{
Gaye, A.
}
\keyword{internal}
