% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySurvDS.R
\name{summarySurvDS}
\alias{summarySurvDS}
\title{Returns summary of survival object.}
\usage{
summarySurvDS(formula = NULL, dataName = NULL)
}
\arguments{
\item{formula}{either NULL or a character string (potentially including '*'
wildcards) specifying a formula.}

\item{dataName}{character string of name of data frame}
}
\value{
a summary of the Cox proportional hazards from the server side environment from the server side environment.
}
\description{
returns a summary of the survival Surv() object from the server side environment.
}
\details{
Serverside aggregate function {coxphSLMADS} called by clientside function.
{ds.summary}.
returns a summary of the survival Surv() object.
This request is not disclosive.
For further details see help for {ds.summary} function.
}
\author{
Soumya Banerjee (2020).
}
