% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitDS.R
\name{survfitDS}
\alias{survfitDS}
\title{Creates a survival survfit object for survival analysis at the serverside environment. 
  This is to be used for eventually plotting survival models.}
\usage{
survfitDS(time = NULL, event = NULL)
}
\arguments{
\item{time}{name of time parameter to be passed to Surv(). 
Should be a character string.}

\item{event}{name of event parameter to be passed to Surv()
Should be character string.}
}
\value{
creates a survfit survival object in the server side environment.
}
\description{
creates a survfit survival object in the server side environment.
}
\details{
Serverside assign function {survfitDS} called by clientside function.
{ds.survfit}.
creates a survfit survival object in the server side environment
This request is not disclosive.
For further details see help for {ds.survfit} function.
}
\author{
Soumya Banerjee (2020).
}
