% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapplyDS.R
\name{tapplyDS}
\alias{tapplyDS}
\title{tapplyDS called by ds.tapply}
\usage{
tapplyDS(X.name, INDEX.names.transmit, FUN.name)
}
\arguments{
\item{X.name, }{the name of the variable to be summarized. Specified
via argument <X.name> of \code{ds.tapply} function}

\item{INDEX.names.transmit, }{the name of a single factor or a vector of names of factors to
index the variable to be summarized. Specified via argument <INDEX.names>
of \code{ds.tapply} function}

\item{FUN.name, }{the name of one of the allowable summarizing functions to be applied.
Specified via argument <FUN.name> of \code{ds.tapply} function.}
}
\value{
an array of the summarized values created by the tapplyDS function. This array
is returned to the clientside. It has the same number of dimensions as INDEX.
}
\description{
Apply one of a selected range of functions to summarize an
outcome variable over one or more indexing factors and write the resultant
summary to the clientside
}
\details{
see details for \code{ds.tapply} function
}
\author{
Paul Burton, Demetris Avraam for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
