% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testObjExistsDS.o.R
\name{testObjExistsDS.o}
\alias{testObjExistsDS.o}
\title{testObjExistsDS.o}
\usage{
testObjExistsDS.o(test.obj.name = NULL)
}
\arguments{
\item{test.obj.name}{a clientside provided character string specifying the variable
whose presence is to be tested in each data source}
}
\description{
The serverside function called by ds.testObjExists.o
}
\details{
Tests whether a given object exists in
all sources. It is called at the end of all
recently written assign functions to check the new (assigned) object has been
created in all sources
}
\author{
Burton PR
}
