% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testObjExistsDS.R
\name{testObjExistsDS}
\alias{testObjExistsDS}
\title{testObjExistsDS}
\usage{
testObjExistsDS(test.obj.name = NULL)
}
\arguments{
\item{test.obj.name}{a client-side provided character string specifying the variable
whose presence is to be tested in each data source}
}
\value{
List with `test.obj.exists` and `test.obj.class`
}
\description{
The server-side function called by ds.testObjExists
}
\details{
Tests whether a given object exists in
all sources. It is called at the end of all
recently written assign functions to check the new (assigned) object has been
created in all sources
}
\author{
Burton PR
}
