% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueDS.R
\name{uniqueDS}
\alias{uniqueDS}
\title{Applies the \code{unique} method to a server-side variable.}
\usage{
uniqueDS(x.name.transmit = NULL)
}
\arguments{
\item{x.name.transmit}{is the name of the variable upon which \code{unique} method will be applied}
}
\value{
the object specified by the \code{newobj} argument
which is written to the server-side.
}
\description{
This function is similar to R function \code{unique}.
}
\details{
The function computes the uniques values of a variable.
}
\author{
Stuart Wheater for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
