% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varDS.R
\name{varDS}
\alias{varDS}
\title{Computes the variance of vector}
\usage{
varDS(x)
}
\arguments{
\item{x}{a character string, the name of a numeric or integer vector}
}
\value{
a list, with the sum of the input variable, the sum of squares of the input variable,
the number of missing values, the number of valid values, the number of total length of the
variable, and a study message indicating whether the number of valid is less than the
disclosure threshold
}
\description{
Calculates the variance.
}
\details{
if the length of input vector is less than the set filter
a missing value is returned.
}
\author{
Amadou Gaye, Demetris Avraam, for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
