% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorDS.R
\name{vectorDS}
\alias{vectorDS}
\title{Creates a vector on the server-side.}
\usage{
vectorDS(...)
}
\arguments{
\item{...}{parameter to be used to form the vector.}
}
\value{
the object specified by the \code{newobj} argument
which is written to the server-side.
}
\description{
This function is similar to R function \code{c}.
}
\details{
The function computes the vectors values.
}
\author{
Stuart Wheater for DataSHIELD Development Team
}
