% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absDS.R
\name{absDS}
\alias{absDS}
\title{Computes the absolute values of the input variable}
\usage{
absDS(x)
}
\arguments{
\item{x}{a string character, the name of a numeric or integer vector}
}
\value{
the object specified by the \code{newobj} argument
of \code{ds.abs} (or default name \code{abs.newobj})
which is written to the serverside. The output object is of class numeric 
or integer.
}
\description{
This function is similar to R function \code{abs}.
}
\details{
The function computes the  absolute values of an input numeric
or integer vector.
}
\author{
Demetris Avraam for DataSHIELD Development Team
}
