% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCharacterDS.R
\name{asCharacterDS}
\alias{asCharacterDS}
\title{Coerces an R object into class character}
\usage{
asCharacterDS(x.name)
}
\arguments{
\item{x.name}{the name of the input object to be coerced to class
character. Must be specified in inverted commas. But this argument is
usually specified directly by \code{x.name} argument of the clientside function
\code{ds.asCharacter}}
}
\value{
the object specified by the \code{newobj} argument (or its default name
"ascharacter.newobj") which is written to the serverside. For further
details see help on the clientside function \code{ds.asCharacter}
}
\description{
this function is based on the native R function \code{as.character}
}
\details{
See help for function \code{as.character} in native R
}
\author{
Amadou Gaye, Paul Burton, Demetris Avraam for DataSHIELD Development Team
}
