% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asFactorDS1.R
\name{asFactorDS1}
\alias{asFactorDS1}
\title{Determines the levels of the input variable in each single study}
\usage{
asFactorDS1(input.var.name = NULL)
}
\arguments{
\item{input.var.name}{the name of the variable that is to be converted to a factor.}
}
\value{
the levels of the input variable.
}
\description{
This function is an aggregate DataSHIELD function that returns the
levels of the input variable from each single study to the client-side function.
}
\details{
The function encodes the input vector as factor and returns its levels in
ascending order if the levels are numerical or in alphabetical order if the levels
are of type character.
}
