% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asListDS.R
\name{asListDS}
\alias{asListDS}
\title{Converts a data.frame or matrix into a list}
\usage{
asListDS(input)
}
\arguments{
\item{input}{the object to turn into list.}
}
\value{
a valid list or an empty list
}
\description{
this function is similar to R function 'as.list'
}
\details{
Unlike the R function 'as.list' some restriction are applied.
}
\author{
Gaye, A.
}
