\name{asMatrixDS}
\alias{asMatrixDS}
\title{Attempts to convert the input into a matrix}
\usage{
asMatrixDS(input)
}
\arguments{
  \item{input}{the object to turn into list.}
}
\value{
a valid list or an empty list
}
\description{
this function is similar to R function 'as.matrix'
}
\details{


Unlike the r function 'as.matrix' some restriction are
applied.
}
\author{
Gaye, A.
}

