% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlotGG_data_Treatment_numericDS.R
\name{boxPlotGG_data_Treatment_numericDS}
\alias{boxPlotGG_data_Treatment_numericDS}
\title{Arrange vector to pass it to the boxplot function}
\usage{
boxPlotGG_data_Treatment_numericDS(vector)
}
\arguments{
\item{vector}{\code{numeric vector} Vector to arrange to be plotted later}
}
\value{
\code{data frame} with the following structure: \cr

 Column 'x': Names on the X axis of the boxplot, aka name of the vector (vector argument) \cr
 Column 'value': Values for that variable \cr
}
\description{
Arrange vector to pass it to the boxplot function
}
