\name{cbindDS}
\alias{cbindDS}
\title{Combines objects by columns}
\usage{
cbindDS(objs)
}
\arguments{
  \item{objs}{a list which contains the names of the
  objects to combine.}
}
\value{
a dataframe
}
\description{
This function is similar to the R base function 'cbind'.
}
\details{


Unlike the R base function 'cbind' this function' output is
a dataframe as that is the type mainly used in DataSHIELD.
}
\author{
Gaye, A.
}

