\name{cDS}
\alias{cDS}
\title{Combines object into a vector or list}
\usage{
cDS(objs)
}
\arguments{
  \item{objs}{a list which contains the names of the
  objects to combine.}
}
\value{
a vector or list
}
\description{
This function is similar to the R base function 'c'.
}
\details{


Unlike the R base function 'c' on vector or list of certain
length are allowed as output
}
\author{
Gaye, A.
}

