% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDS.R
\name{corDS}
\alias{corDS}
\title{Computes correlation between two or more vectors}
\usage{
corDS(x = NULL, y = NULL, use = NULL)
}
\arguments{
\item{x}{a character, the name of a vector, dataframe or matrix}

\item{y}{(optional) a character, the name of a vector, dataframe or matrix}

\item{use}{a character string giving a method for computing covariances in the 
presence of missing values. This must be one of the strings: "everything", "all.obs", 
"complete.obs", "na.or.complete", or "pairwise.complete.obs".}
}
\value{
correlation
}
\description{
this function is similar to R function 'cor'
}
\author{
Gaye, A.
}
