\name{corDS}
\alias{corDS}
\title{Computes correlation between two or more vectors}
\usage{
corDS(x = NULL, y = NULL)
}
\arguments{
  \item{x}{a character, the name of a vector, dataframe or
  matrix}

  \item{y}{(optional) a character, the name of a vector,
  dataframe or matrix}
}
\value{
correlation
}
\description{
this function is similar to R function 'cor'
}
\details{

}
\author{
Gaye, A.
}

