% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTestDS.R
\name{corTestDS}
\alias{corTestDS}
\title{Tests for correlation between paired samples}
\usage{
corTestDS(x, y)
}
\arguments{
\item{x}{a character string providing  the name of a numerical vector.}

\item{y}{a character string providing  the name of a numerical vector.}
}
\value{
the results of the Pearson test.
}
\description{
This function is similar to R function \code{cor.test}.
}
\details{
The function runs a two-sided Pearson test with a 0.95 confidence level.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
