\name{covDS}
\alias{covDS}
\title{Computes covariance between two or more vectors}
\usage{
covDS(x = NULL, y = NULL)
}
\arguments{
  \item{x}{a character, the name of a vector, dataframe or
  matrix}

  \item{y}{(optional) a character, the name of a vector,
  dataframe or matrix}
}
\value{
covariance
}
\description{
this function is similar to R function 'cov'
}
\details{

}
\author{
Gaye, A.
}

