% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphSLMADS.R
\name{coxphSLMADS}
\alias{coxphSLMADS}
\title{Performs survival analysis using the Cox proportional hazards model at the serverside environment.}
\usage{
coxphSLMADS(formula = NULL, dataName = NULL, weights = NULL)
}
\arguments{
\item{formula}{either NULL or a character string (potentially including '*'
wildcards) specifying a formula.}

\item{dataName}{character string of name of data frame}

\item{weights}{vector of case weights}
}
\value{
a summary of the Cox proportional hazards from the server side environment from the server side environment.
}
\description{
returns a summary of the Cox proportional hazards from the server side environment.
}
\details{
Serverside aggregate function {coxphSLMADS} called by clientside function.
{ds.coxphSLMA}.
returns a summary of the Cox proportional hazards from the server side environment from the server side environment.
This request is not disclosive as it only returns a string.
For further details see help for {ds.coxphSLMA} function.
}
\author{
Soumya Banerjee and Tom Bishop (2020).
}
