% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphSLMADS.R
\name{coxphSLMADS}
\alias{coxphSLMADS}
\title{returns a string from a serverside environment}
\usage{
coxphSLMADS(search.filter = NULL)
}
\arguments{
\item{search.filter}{either NULL or a character string (potentially including '*'
wildcards) specifying required search criteria. This argument is
fully specified by its corresponding argument in the clientside function.}
}
\value{
a string containing a Hello World from server-side appended to whatever string was passed from client-side
}
\description{
returns a string from the server side environment afetr appending it to a string passed from the client side
}
\details{
Serverside aggregate function {retStrDS} called by clientside function.
{ds.retStr}.
returns a string from the server side environment afetr appending it to a string passed from the client side. 
This request is not disclosive as it only returns a string.
For further details see help for {ds.retStr} function.
}
\author{
Soumya Banerjee (2020).
}
