% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimDS.R
\name{dimDS}
\alias{dimDS}
\title{Returns the dimension of a data frame or matrix}
\usage{
dimDS(x)
}
\arguments{
\item{x}{a string character, the name of a dataframe or matrix}
}
\value{
a list with two elements: \code{dim} (the dimension of the input object)
  and \code{class} (the class of the input object, for client-side consistency checking)
}
\description{
This function is similar to R function \code{dim}.
}
\details{
The function returns the dimension of the input dataframe or matrix
}
\author{
Demetris Avraam, for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
