\name{dimDS}
\alias{dimDS}
\title{Returns the dimensions of a table structure}
\usage{
dimDS(x)
}
\arguments{
  \item{x}{a character, the name of a matrix or a
  dataframe.}
}
\value{
the number of rows and the number of columns
}
\description{
Returns the number of rows and the number of columns.
}
\details{


This is similar to R base function \code{dim} but here the
input is a character.
}
\author{
Gaye, A.
}

