% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsBase-package.R
\docType{package}
\name{dsBase-package}
\alias{dsBase}
\alias{dsBase-package}
\title{dsBase: 'DataSHIELD' Server Side Base Functions}
\description{
Base 'DataSHIELD' functions for the server side. 'DataSHIELD' is a software package which allows you to do non-disclosive federated analysis on sensitive data. 'DataSHIELD' analytic functions have been designed to only share non disclosive summary statistics, with built in automated output checking based on statistical disclosure control. With data sites setting the threshold values for the automated output checks. For more details, see 'citation("dsBase")'.
}
\author{
\strong{Maintainer}: Stuart Wheater \email{stuart.wheater@arjuna.com} (\href{https://orcid.org/0009-0003-2419-1964}{ORCID})

Authors:
\itemize{
  \item Paul Burton (\href{https://orcid.org/0000-0001-5799-9634}{ORCID})
  \item Rebecca Wilson (\href{https://orcid.org/0000-0003-2294-593X}{ORCID})
  \item Olly Butters (\href{https://orcid.org/0000-0003-0354-8461}{ORCID})
  \item Patricia Ryser-Welch (\href{https://orcid.org/0000-0002-0070-0264}{ORCID})
  \item Alex Westerberg
  \item Leire Abarrategui
  \item Roberto Villegas-Diaz (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})
  \item Demetris Avraam (\href{https://orcid.org/0000-0001-8908-2441}{ORCID})
  \item Yannick Marcon \email{yannick.marcon@obiba.org} (\href{https://orcid.org/0000-0003-0138-2023}{ORCID})
  \item Tom Bishop
  \item Amadou Gaye (\href{https://orcid.org/0000-0002-1180-2792}{ORCID})
  \item Xavier Escribà-Montagut (\href{https://orcid.org/0000-0003-2888-8948}{ORCID})
}

}
\keyword{internal}
