% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDS1.R
\name{glmDS1}
\alias{glmDS1}
\title{glmDS1 called by ds.glm}
\usage{
glmDS1(formula, family, weights, offset, data)
}
\arguments{
\item{formula}{a glm() formula consistent with R syntax eg U~x+y+Z to regress
variables U on x,y and Z}

\item{family}{a glm() family consistent with R syntax eg "gaussian", "poisson",
"binomial"}

\item{weights}{an optional variable providing regression weights}

\item{offset}{the offset}

\item{data}{an
optional character string specifying a data.frame object holding the data to be
analysed under the specified model}
}
\value{
List with values from GLM model.
}
\description{
This is the first server-side aggregate function called by ds.glm
}
\details{
It is an
aggregation function that sets up the model structure
and creates the starting beta.vector that feeds, via ds.glm, into glmDS2 to enable
iterative fitting of the generalized linear model that has been
been specified. For more details please see the extensive header for ds.glm.
}
\author{
Burton PR for DataSHIELD Development Team
}
