% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNaDS.R
\name{isNaDS}
\alias{isNaDS}
\title{Checks if a vector is empty}
\usage{
isNaDS(x)
}
\arguments{
\item{x}{a character string, the name of a server-side vector}
}
\value{
TRUE if the vector contains only NAs, FALSE otherwise
}
\description{
this function is similar to R function \code{is.na} but instead of a vector
of booleans it returns just one boolean to tell if all the element are missing values.
}
\author{
Gaye, A.
}
