% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidDS.R
\name{isValidDS}
\alias{isValidDS}
\title{Checks if an input is valid}
\usage{
isValidDS(obj)
}
\arguments{
\item{obj, }{a vector (numeric, integer, factor, character), data.frame or matrix}
}
\value{
a boolean, TRUE if input is valid or FALSE if not.
}
\description{
Tells if an object on the server side is valid.
}
\details{
This function checks if an object is valid.
}
\author{
Gaye, A.
}
