% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthDS.R
\name{lengthDS}
\alias{lengthDS}
\title{Returns the length of a vector or list}
\usage{
lengthDS(x)
}
\arguments{
\item{x}{a string character, the name of a vector or list}
}
\value{
a list with two elements: \code{length} (the number of elements of the input
  vector or list) and \code{class} (the class of the input object, for client-side
  consistency checking)
}
\description{
This function is similar to R function \code{length}.
}
\details{
The function returns the length of the input vector or list.
}
\author{
Demetris Avraam, for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
