% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelsDS.R
\name{levelsDS}
\alias{levelsDS}
\title{Returns the levels of a factor vector}
\usage{
levelsDS(x)
}
\arguments{
\item{x}{a factor vector}
}
\value{
a list with two elements: \code{Levels} (the factor levels present
  in the vector) and \code{class} (the class of the input object, for
  client-side consistency checking)
}
\description{
This function is similar to R function \code{levels}.
}
\details{
The function returns the levels of the input vector or list.
}
\author{
Alex Westerberg, for DataSHIELD Development Team

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
