% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelsDS.R
\name{levelsDS}
\alias{levelsDS}
\title{Returns the levels of a factor vector}
\usage{
levelsDS(x)
}
\arguments{
\item{x}{a factor vector}
}
\value{
a list, the factor levels present in the vector
}
\description{
This function is similar to R function \code{levels}.
}
\details{
The function returns the levels of the input vector or list.
}
\author{
Alex Westerberg, for DataSHIELD Development Team
}
