% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisDS1.R
\name{lexisDS1}
\alias{lexisDS1}
\title{lexisDS1}
\usage{
lexisDS1(exitCol = NULL)
}
\arguments{
\item{exitCol}{a character string specifying the variable holding the time that each individual is censored or fails}
}
\value{
List with `max.time`
}
\description{
The first server-side function called by ds.lexis.
}
\details{
This is an aggregate function.
For more details see the extensive header for ds.lexis.
}
\author{
Burton PR
}
