\name{listDS}
\alias{listDS}
\title{Coerce objects into a list}
\usage{
listDS(input = NULL, eltnames = NULL)
}
\arguments{
  \item{input}{a list of objects to coerce into a list}

  \item{eltnames}{a character list, the names of the
  elements in the list.}
}
\value{
a list
}
\description{
this function is similar to R function 'list'
}
\details{


Unlike the R function 'list' it takes also a vector of
characters, the names of the elements in the output list.
}
\author{
Gaye, A.
}

