% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixInvertDS.R
\name{matrixInvertDS}
\alias{matrixInvertDS}
\title{matrixInvertDS serverside assign function called by ds.matrixInvert}
\usage{
matrixInvertDS(M1.name = NULL)
}
\arguments{
\item{M1.name}{A character string specifying the name of the matrix to be inverted}
}
\value{
Output is the matrix representing the inverse of A which is written
to the serverside. For more details see help for ds.matrixInvert
}
\description{
Inverts a square matrix A and writes the output to the serverside
}
\details{
Undertakes standard matrix inversion. This operation is only
possible if the number of columns and rows of A are the same and the matrix
is non-singular - positive definite (eg there is no row or column that is all zeros)
}
\author{
Paul Burton for DataSHIELD Development Team
}
