% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTransposeDS.R
\name{matrixTransposeDS}
\alias{matrixTransposeDS}
\title{matrixTransposeDS serverside assign function called by ds.matrixTranspose}
\usage{
matrixTransposeDS(M1.name = NULL)
}
\arguments{
\item{M1.name}{A character string specifying the name of the matrix to be transposed}
}
\value{
Output is the matrix representing the transpose of A which is written
to the serverside. For more details see help for ds.matrixTranspose
}
\description{
Transposes a matrix A and writes the output to the serverside
}
\details{
Undertakes standard matrix transposition. This operation converts matrix
A to matrix C where element C[i,j] of matrix C equals element A[j,i] of matrix
A. Matrix A therefore has the same number of rows as matrix C has columns and
vice versa.
}
\author{
Paul Burton for DataSHIELD Development Team
}
