% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanDS.R
\name{meanDS}
\alias{meanDS}
\title{Computes statistical mean of a vectores}
\usage{
meanDS(xvect)
}
\arguments{
\item{xvect}{a vector}
}
\value{
a numeric, the statistical mean
}
\description{
Calculates the mean value.
}
\details{
if the length of input vector is less than the set filter
a missing value is returned.
}
\author{
Gaye, A.
}
