% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSdGpDS.R
\name{meanSdGpDS}
\alias{meanSdGpDS}
\title{MeanSdGpDS}
\usage{
meanSdGpDS(X, INDEX)
}
\arguments{
\item{X}{a clientside supplied character string identifying the variable for which
means/SDs are to be calculated}

\item{INDEX}{a clientside supplied character string identifying the factor across
which means/SDs are to be calculated}
}
\description{
Serverside function called by ds.meanSdGp
}
\details{
Computes the mean and standard deviation across groups defined by one
factor
}
\author{
Burton PR
}
