% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataDS.R
\name{metadataDS}
\alias{metadataDS}
\title{Returns the metadata, if any, about the specified variable}
\usage{
metadataDS(x)
}
\arguments{
\item{x}{a string character, containing the name  of the specified variable}
}
\value{
a list containing the metadata. The elements of the list will depend
on the meatadata available.
}
\description{
This function returns metadata, if any, about specified variable.
}
\details{
The function returns the metadata, obtained from attributes function.
}
\author{
Stuart Wheater, for DataSHIELD Development Team
}
