\name{namesDS}
\alias{namesDS}
\title{Returns the names of a list}
\usage{
namesDS(xlist)
}
\arguments{
  \item{xlist}{a list}
}
\value{
a character vector or NULL if the list does not have names
}
\description{
Returns the names of the object in the list.
}
\details{


This is similar to R base function \code{names} but
restricted to list types only.
}
\author{
Gaye, A.
}

