% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numNaDS.R
\name{numNaDS}
\alias{numNaDS}
\title{Counts the number of missing values}
\usage{
numNaDS(x)
}
\arguments{
\item{x}{a character string, the name of a server-side vector}
}
\value{
a list with two elements: \code{numNA} (an integer, the number of
  missing values) and \code{class} (the class of the input object, for
  client-side consistency checking)
}
\description{
this function just counts the number of missing entries
in a vector.
}
\author{
Gaye, A.

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
