% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numNaDS.R
\name{numNaDS}
\alias{numNaDS}
\title{Counts the number of missing values}
\usage{
numNaDS(xvect)
}
\arguments{
\item{xvect}{a vector}
}
\value{
an integer, the number of missing values
}
\description{
this function just counts the number of missing entries 
in a vector.
}
\author{
Gaye, A.
}
