% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileMeanDS.R
\name{quantileMeanDS}
\alias{quantileMeanDS}
\title{Generates quantiles and mean information without maximum and minimum}
\usage{
quantileMeanDS(x)
}
\arguments{
\item{x}{a character string, the name of a numeric or integer vector}
}
\value{
a list, with \code{quantiles}, a numeric vector that represents the sample
quantiles, and \code{class}, the class of the input object for client-side consistency
checking
}
\description{
the probabilities 5%, 10%, 25%, 50%, 75%, 90%, 95% and the mean
are used to compute the corresponding quantiles.
}
\author{
Burton, P.; Gaye, A.

Tim Cadman, Genomics Coordination Centre, UMCG, Netherlands
}
