\name{rangeDS}
\alias{rangeDS}
\title{returns the minimum and maximum of a numeric vector}
\usage{
rangeDS(xvect)
}
\arguments{
  \item{xvect}{a numerical}
}
\value{
a numeric vector which contains the minimum and the maximum
}
\description{
this function is similar to R function \code{range} but
instead to not return the real minimum and maximum, the
computed values are multiplied by a very small random
number.
}
\details{

}
\author{
Gaye, A.
}

