% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeDS.R
\name{rangeDS}
\alias{rangeDS}
\title{returns the minimum and maximum of a numeric vector}
\usage{
rangeDS(xvect)
}
\arguments{
\item{xvect}{a numerical}
}
\value{
a numeric vector which contains the minimum and the maximum values of the vector
}
\description{
this function is similar to R function \code{range} but instead to not return
the real minimum and maximum, the computed values are multiplied by a very small random number.
}
\author{
Amadou Gaye, Demetris Avraam for DataSHIELD Development Team
}
