% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recodeLevelsDS.R
\name{recodeLevelsDS}
\alias{recodeLevelsDS}
\title{Recodes the levels of a categorical variables}
\usage{
recodeLevelsDS(x = NULL, classes = NULL)
}
\arguments{
\item{x}{a factor vector}

\item{classes}{a character vector the levels of the  newt factor vector}
}
\value{
a factor vector with the new levels
}
\description{
The functions uses the input factor and generates a new factor
with new levels.
}
\details{

}
\author{
Gaye, A.
}

