% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeValuesDS.R
\name{recodeValuesDS}
\alias{recodeValuesDS}
\title{recodeValuesDS an assign function called by ds.recodeValues}
\usage{
recodeValuesDS(
  var.name.text = NULL,
  values2replace.text = NULL,
  new.values.text = NULL,
  missing = NULL
)
}
\arguments{
\item{var.name.text}{a character string providing the name for the vector representing the
variable to be recoded. <var.name.text> argument generated and passed directly to
recodeValuesDS by ds.recodeValues}

\item{values2replace.text}{a character string specifying the values in the
vector specified by the argument <var.name.text> that are to be replaced by new
values as specified in the new.values.vector. The <values2replace.text> argument
is generated and passed directly to recodeValuesDS by ds.recodeValues. In effect, the
<values2replace.vector> argument of the ds.recodeValues function is converted
to a character string format that is acceptable to the DataSHIELD R parser in the data repository
and so can be accepted by recodeValuesDS}

\item{new.values.text}{a character string specifying the new values to which
the specified values in the vector <var.name> are to be converted.
The <new.values.text> argument is generated and passed directly to recodeValuesDS
by ds.recodeValues. In effect, the <new.values.vector> argument of the
ds.recodeValues function is converted to a character string format that is
acceptable to the DataSHIELD R parser in the data repository
and so can be used in the call to recodeValuesDS.}

\item{missing}{if supplied, any missing values in the variable referred to by var.name.text 
will be replaced by this value.}
}
\value{
the object specified by the <newobj> argument (or default name '<var.name>_recoded')
initially specified in calling ds.recodeValues. The output object (the required
recoded variable called <newobj> is written to the serverside.
}
\description{
This function recodes specified values of elements in a vector into
a matched set of alternative specified values.
}
\details{
For all details see the help header for ds.recodeValues
}
\author{
Paul Burton, Demetris Avraam for DataSHIELD Development Team
}
