% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFilterDS.R
\name{setFilterDS}
\alias{setFilterDS}
\title{Sets the privacy level}
\usage{
setFilterDS(x = getOption("datashield.privacyLevel", default = 5))
}
\arguments{
\item{x}{a dummy argument}
}
\value{
an integer between 1 and 5
}
\description{
This is an internal function that set the number of observation
that are allowed in any table or vector. The function gets the value from opal
and if no value was set in opal it uses the default value 5.
}
\details{
The function is called by server side functios
}
\author{
Gaye, A.
}
\keyword{internal}
