% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewnessDS2.R
\name{skewnessDS2}
\alias{skewnessDS2}
\title{Calculates the skewness of a numeric variable}
\usage{
skewnessDS2(x, global.mean)
}
\arguments{
\item{x}{a string character, the name of a numeric variable.}

\item{global.mean}{a numeric, the combined mean of the input variable across all studies.}
}
\value{
a list including the sum of cubes of variable x, the sum of squares of variable x, the number of
valid observations and a validity message.
}
\description{
This function calculates summary statistics that are returned to the client-side and 
used for the estimation of the combined skewness of a numeric variable across all studies.
}
\details{
The function calculates the sum of squares, the sum of cubes and the number of valid observations
of the input variable x.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
