% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtDS.R
\name{sqrtDS}
\alias{sqrtDS}
\title{Computes the square root values of the input variable}
\usage{
sqrtDS(x)
}
\arguments{
\item{x}{a string character, the name of a numeric or integer vector}
}
\value{
the object specified by the \code{newobj} argument
of \code{ds.sqrt} (or default name \code{sqrt.newobj})
which is written to the server-side. The output object is of class numeric 
or integer.
}
\description{
This function is similar to R function \code{sqrt}.
}
\details{
The function computes the square root values of an input numeric
or integer vector.
}
\author{
Demetris Avraam for DataSHIELD Development Team
}
