% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvDS.R
\name{SurvDS}
\alias{SurvDS}
\title{Creates a survival object for survival analysis using the Cox proportional hazards model at the serverside environment}
\usage{
SurvDS(start = NULL, stop = NULL, event = NULL)
}
\arguments{
\item{start}{name of start time parameter to be passed to Surv(). 
Should be a character string.}

\item{stop}{name of stop time parameter to be passed to Surv(). 
Should be a character string.}

\item{event}{name of event parameter to be passed to Surv()
Should be character string.}
}
\value{
a survival::Surv() object from the server side environment.
}
\description{
returns a summary of the Cox proportional hazards from the server side environment.
}
\details{
Serverside assign function {SurvDS} called by clientside function.
{ds.Surv}.
returns a Survival object for use in Cox proportional hazards from the server side environment from the server side environment.
This request is not disclosive as it only returns a string.
For further details see help for {ds.Surv} function.
}
\author{
Soumya Banerjee and Tom Bishop (2020).
}
