% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvDS.R
\name{SurvDS}
\alias{SurvDS}
\title{Creates a survival object for survival analysis using the Cox proportional hazards model at the serverside environment}
\usage{
SurvDS(formula = NULL, dataName = NULL)
}
\arguments{
\item{formula}{either NULL or a character string (potentially including '*'
wildcards) specifying a formula.}

\item{dataName}{character string of name of data frame}
}
\value{
a summary of the Cox proportional hazards from the server side environment from the server side environment.
}
\description{
returns a summary of the Cox proportional hazards from the server side environment.
}
\details{
Serverside assign function {SurvDS} called by clientside function.
{ds.Surv}.
returns a Survival object for use in Cox proportional hazards from the server side environment from the server side environment.
This request is not disclosive as it only returns a string.
For further details see help for {ds.Surv} function.
}
\author{
Soumya Banerjee (2020).
}
