% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableDS.R
\name{tableDS}
\alias{tableDS}
\title{tableDS is the first of two serverside aggregate functions
called by ds.table}
\usage{
tableDS(
  rvar.transmit,
  cvar.transmit,
  stvar.transmit,
  rvar.all.unique.levels.transmit,
  cvar.all.unique.levels.transmit,
  stvar.all.unique.levels.transmit,
  exclude.transmit,
  useNA.transmit,
  force.nfilter.transmit
)
}
\arguments{
\item{rvar.transmit}{is a character string (in inverted commas) specifiying the
name of the variable defining the rows in all of the 2 dimensional
tables that form the output. Fully specified by <rvar> argument in {ds.table}.
For more information see help for {ds.table}}

\item{cvar.transmit}{is a character string specifiying the
name of the variable defining the columns in all of the 2 dimensional
tables that form the output. Fully specified by <cvar> argument in {ds.table}.
For more information see help for {ds.table}}

\item{stvar.transmit}{is a character string specifiying the
name of the variable that indexes the separate two dimensional
tables in the output if the call specifies a 3 dimensional table.
Fully specified by <stvar> argument in {ds.table}.
For more information see help for {ds.table}}

\item{rvar.all.unique.levels.transmit}{is a character string containing all
unique level in rvar, across the studies, separated by ','.}

\item{cvar.all.unique.levels.transmit}{is a character string containing all
unique level in cvar, across the studies, separated by ','.}

\item{stvar.all.unique.levels.transmit}{is a character string containing all
unique level in stvar, across the studies, separated by ','.}

\item{exclude.transmit}{for information see help on <exclude> argument
of {ds.table}. Fully specified by <exclude> argument of {ds.table}}

\item{useNA.transmit}{for information see help on <useNA> argument
of {ds.table}. Fully specified by <useNA> argument of {ds.table}}

\item{force.nfilter.transmit}{for information see help on <force.nfilter> argument 
of {ds.table}. Fully specified by <force.nfilter> argument of {ds.table}}
}
\value{
For information see help for {ds.table}
}
\description{
creates 1-dimensional, 2-dimensional and 3-dimensional
tables using the {table} function in native R.
}
\details{
this serverside function is the workhorse of {ds.table} - creating
the table requested in the format specified by {ds.table}. For more
information see help for {ds.table} in DataSHIELD and the {table} function
in native R.
}
\author{
Paul Burton for DataSHIELD Development Team, 13/11/2019
}
