% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uniqueDS.R
\name{uniqueDS}
\alias{uniqueDS}
\title{Returns unique rows of a data frame based on selected columns}
\usage{
uniqueDS(x, f)
}
\arguments{
\item{x}{a data frame}

\item{f}{a vector of column names for checking uniqueness}
}
\value{
all columns, but only rows unique in the selected columns
}
\description{
Returns unique rows of a data frame based on selected columns
}
\details{


Needs checking to see if it will work with a matrix. All columns are returne
For columns outside the selection, the first value is taken
}
\author{
Bishop, T.
}

