\name{varDS}
\alias{varDS}
\title{Computes the variance of vector with more than 4 entries}
\usage{
varDS(xvect)
}
\arguments{
  \item{xvect}{a vector}
}
\value{
a numeric, the variance
}
\description{
Calculates the variance.
}
\details{


if the length of input vector is less than the set filter a
missing value is returned.
}
\author{
Gaye, A.
}

